requirements.txt
asgiref==3.3.0
attrs==20.3.0
coverage==5.3
Django==3.1.3
django-cors-headers==3.5.0
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.2
gunicorn==20.0.4
iniconfig==1.1.1
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
pytest-cov==2.10.1
pytest-django==4.1.0
pytest-pythonpath==0.7.3
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
